requirements.txt
Babel==2.3.4
configparser==3.5.0
debtcollector==1.5.0
iso8601==0.1.11
keystoneauth1==2.9.0
monotonic==1.1
msgpack-python==0.4.7
netaddr==0.7.18
netifaces==0.10.4
oslo.i18n==3.7.0
oslo.serialization==2.10.0
oslo.utils==3.14.0
pbr==1.10.0
positional==1.1.1
prettytable==0.7.2
python-novaclient==4.1.0
pytz==2016.4
requests==2.10.0
simplejson==3.8.2
six==1.10.0
stevedore==1.15.0
wrapt==1.10.8
